requirements.txt
pycairo==1.20.0